modular_dash/requirements.txt
plotly
pandas
shiny
requests
zeep
shinywidgets
pins
pyarrow
python-dotenv